Create a single top-level li file.
