+++ pkgs/cupertino_http/example/integration_test/client_conformance_test.dart
@@ -2,7 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:cupertino_http/cupertino_client.dart';
 import 'package:cupertino_http/cupertino_http.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http_client_conformance_tests/http_client_conformance_tests.dart';
